// libs/features/establishment/src/lib/shared/enums/establishment-eligibility-enum.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

export enum EstablishmentEligibilityEnum {
  CBM = 'Branch To Main',
  DELINK = 'Delink an Establishment',
  DELINK_ONLY = 'Delink Only Eligibility',
  CLOSE_EST = 'Terminate Establishment'
}